Created auth.

+++ lib/Models/preferences.dart
@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 class PreferencesModel extends ChangeNotifier {
   static List<String> _dietTypes = [];
   static List<String> _restrictions = [];
-  static double _maxPrepTime = -1; // any
+  static int _maxPrepTime = -1; // any
 
 
   List<String> get dietTypes => _dietTypes;
   List<String> get restrictions => _restrictions;
-  double get maxPrepTime => _maxPrepTime;
+  int get maxPrepTime => _maxPrepTime;
 
   set dietTypes(List<String> dietTypes) {
     _dietTypes = dietTypes;
@@ -18,7 +18,7 @@ class PreferencesModel extends ChangeNotifier {
     _restrictions = restrictions;
     notifyListeners();
   }
-  set maxPrepTime(double maxPrepTime) {
+  set maxPrepTime(int maxPrepTime) {
     _maxPrepTime = maxPrepTime;
     notifyListeners();
   }
